add voters

+++ src/Controller/Front/AdvertController.php
@@ -6,6 +6,7 @@ namespace App\Controller\Front;
 use App\Entity\Advert;
 use App\Entity\GamePlatform;
 use App\Form\Front\AdvertType;
+use App\Security\AdvertVoter;
 use Doctrine\Common\Persistence\ObjectManager;
 use Doctrine\ORM\EntityManager;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
@@ -70,6 +71,7 @@ class AdvertController extends AbstractController
 	 * @return \Symfony\Component\HttpFoundation\RedirectResponse
 	 */
 	public function edit(Request $request, Advert $advert, ObjectManager $em): Response {
+		$this->denyAccessUnlessGranted(AdvertVoter::OWNER, $advert);
 		$form = $this->createForm(AdvertType::class, $advert);
 		$form->handleRequest($request);
 		if($form->isSubmitted() && $form->isValid()){
@@ -88,6 +90,7 @@ class AdvertController extends AbstractController
 	 * @return \Symfony\Component\HttpFoundation\RedirectResponse
 	 */
 	public function delete(Request $request, Advert $advert) {
+		$this->denyAccessUnlessGranted(AdvertVoter::OWNER, $advert);
 		if ($this->isCsrfTokenValid('delete'.$advert->getId(), $request->request->get('_token'))) {
 			$entityManager = $this->getDoctrine()->getManager();
 			$entityManager->remove($advert);
